Restructure logging (#1)
* restructured logging

* fixed build errors

* include code to protect against injections

* fixed issue in the signals that was causing process to hang

* fixed an incompatibility with the v1.4 of libflac

* Inconsistent signature errors fix

* Fixed some conflicting types issues, and apparently lame deprecated some of its interface, so had to do some replacements

* Cleaned up a baffling(and I suspect never used) section of code

* Updated the autotools stuff to clean up the deprecation warnings there

* Replaced all references to RETSIGTYPE with void

+++ src/in_vorbis.c
@@ -44,7 +44,7 @@ typedef struct {
 } ices_vorbis_in_t;
 
 /* -- static prototypes -- */
-static int ices_vorbis_readpcm(input_stream_t* self, size_t len,
+static ssize_t ices_vorbis_readpcm(input_stream_t* self, size_t len,
 			       int16_t* left, int16_t* right);
 static int ices_vorbis_close(input_stream_t* self);
 static void in_vorbis_parse(input_stream_t* self);
@@ -131,7 +131,7 @@ int ices_vorbis_open(input_stream_t* self, char* buf, size_t len) {
 	return 0;
 }
 
-static int ices_vorbis_readpcm(input_stream_t* self, size_t olen, int16_t* left,
+static ssize_t ices_vorbis_readpcm(input_stream_t* self, size_t olen, int16_t* left,
 			       int16_t* right) {
 	ices_vorbis_in_t* vorbis_data = (ices_vorbis_in_t*) self->data;
 	int link;
